Found the logging rate problem, but not the solution.
